fixed a problem about first build file deletion in case is js

+++ index.js
@@ -11,18 +11,20 @@ function malta_es6(o, options) {
 		outname = o.name.replace(/\.(ts|coffee)$/, '.js'),
 		pluginName = path.basename(path.dirname(__filename)),
 		args = [o.name, '-o', outname];
+
 	options = options || {};
 	if ('plugins' in options) args.push('--plugins', options.plugins);
 	if ('presets' in options) args.push('--presets', options.presets);
 
 	return function (solve, reject){
+		var doDelete = !!(o.name.match(/\.(ts|coffee)$/));
 		try {
 			var ls = child_process.spawn('babel', args);
 			ls.on('exit', function (code) {
 				if (code == 0) {
 					o.content = fs.readFileSync(outname) + "";
 					msg = 'plugin ' + pluginName.white() + ' wrote ' + outname;
-					fs.unlink(o.name);
+					doDelete && fs.unlink(o.name);
 					solve(o);
 					self.notifyAndUnlock(start, msg);
 				}
@@ -31,7 +33,7 @@ function malta_es6(o, options) {
 				console.log("ERROR".red());
 				msg = 'plugin ' + pluginName.white() + ' compilation error';
 				console.log((err+"").white());
-				fs.unlink(o.name);
+				doDelete && fs.unlink(o.name);
 				solve(o);
 				self.notifyAndUnlock(start, msg);
 			});
